Make sure generated files are always generated and stored in the source JAR
- Also fix a generation issue on Windows

Fixes #281

+++ api/docs/org.jboss.tools.rsp.schema/src/main/java/org/jboss/tools/rsp/api/schema/JSONUtility.java
@@ -34,8 +34,10 @@ public class JSONUtility {
 	public void cleanFolder() {
 		Path folder = getDaoJsonFolder();
 		File[] jsons = folder.toFile().listFiles();
-		for( int i = 0; i < jsons.length; i++ ) {
-			jsons[i].delete();
+		if (jsons != null) {
+			for( int i = 0; i < jsons.length; i++ ) {
+				jsons[i].delete();
+			}
 		}
 	}
